MidiMapper refactor and changes
MidiMapper has greatly simplified. Instead of using an tensorflow queue
to manage the enqueing of the embeddings, it now holds the embeddigns
inside a python queue as a queue of arrays. each of these arrays is
initialised as a numpy zeros of shape = (lc_embeddings) and then the
upsampler algorithm sets them to 1 according to embedding strategy. This
removes the reliance of mapper on tensorflow session and coordinator,
reducing complexity. It may also result in removal of subtle bugs or
efficiency gains.